+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,6 @@
-import java.io.IOException;
+package duke;
+
+import java.lang.Exception;
 
 public class DukeException extends Exception {
     public DukeException(String str) {
