+++ src/main/java/duke/Events.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Events extends Task {
     private String strtime;
     private String endtime;
