+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.Task;
+
 import java.io.Serializable;
 import java.util.ArrayList;
