+++ src/main/java/duke/Ui.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
