+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,11 @@
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Task;
+import duke.TaskList;
+import duke.Ui;
+
 public class Command {
 
     public void execute(TaskList list, Storage store, Ui ui) throws DukeException {
