+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Task;
+import duke.TaskList;
+import duke.Ui;
+
 public class DeleteCommand extends Command {
 
     private int deleteNumber;
