+++ src/main/java/duke/commands/TodoCommand.java
@@ -1,29 +1,26 @@
-public class EventCommand extends Command {
-    private String taskName;
-    private String sTime;
-    private String eTime;
+package duke.commands;
 
-    public EventCommand(String taskName, String sTime, String eTime) {
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Todos;
+import duke.Ui;
+
+public class TodoCommand extends Command {
+
+    private String taskName;
+
+    public TodoCommand(String taskName) {
         this.taskName = taskName;
-        this.sTime = sTime;
-        this.eTime = eTime;
     }
 
     public String getTaskName() {
         return this.taskName;
     }
 
-    public String getStartTime() {
-        return this.sTime;
-    }
-
-    public String getEndTime() {
-        return this.eTime;
-    }
-
     @Override
     public void execute(TaskList list, Storage store, Ui ui) throws DukeException {
-        list.add(new Events(taskName, sTime, eTime));
+        list.add(new Todos((taskName)));
         store.save(list);
         ui.showMessage("Got it. I've added this task:");
         ui.showMessage(list.get(list.size() - 1).toString());
